Move Scale menu to OverViewMenusImpl
