+++ core/interfaces/src/main/kotlin/app/aaps/core/interfaces/overview/OverviewMenus.kt
@@ -1,5 +1,6 @@
 package app.aaps.core.interfaces.overview
 
+import android.widget.Button
 import android.widget.ImageButton
 
 interface OverviewMenus {
@@ -22,7 +23,8 @@ interface OverviewMenus {
 
     val setting: List<Array<Boolean>>
     fun loadGraphConfig()
-    fun setupChartMenu(chartButton: ImageButton)
+    fun setupChartMenu(chartButton: ImageButton, scaleButton: Button)
     fun enabledTypes(graph: Int): String
     fun isEnabledIn(type: CharType): Int
+    fun scaleString(rangeToDisplay: Int): String
 }
